# Backend/MainService/openapi.yaml
openapi: 3.1.0
info:
  title: User Service API
  version: 1.0.0
  description: OpenAPI спецификация для основного сервиса
paths:
...
  /api/v1/main_service/articles:
    post:
      tags:
        - Article
      summary: Создание статьи
      description: Создает новую статью от имени текущего пользователя
      operationId: createArticle
      parameters:
        - $ref: '#/components/parameters/XUserNameHeader'
      requestBody:
        $ref: '#/components/requestBodies/ArticleRequestWithImage'
      responses:
        '201':
          description: Статья была успешно создана
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArticleFullResponse'
        '400':
          description: Ошибка валидации
        '401':
          description: Неавторизованный доступ
    get:
      tags:
        - Article
      summary: Получение списка статей
      description: Возвращает список статей с фильтрацией и пагинацией
      operationId: getArticles
      parameters:
        - in: query
          name: category
          schema:
            type: string
          description: Фильтр по категории
        - in: query
          name: query
          schema:
            type: string
          description: Поисковый запрос по заголовку или содержимому
        - in: query
          name: moderationPassed
          schema:
            type: boolean
          description: Фильтр по статусу модерации
        - in: query
          name: page
          schema:
            type: integer
            default: 0
            minimum: 0
          description: Номер страницы для пагинации
        - in: query
          name: size
          schema:
            type: integer
            default: 20
            minimum: 1
          description: Размер страницы для пагинации
      responses:
        '200':
          description: Список статей
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArticlePageResponse'
        '400':
          description: Ошибка параметров запроса
...
  /api/v1/main_service/articles/me:
    get:
      tags:
        - Article
      summary: Получение статей созданных текущим пользователем с пагинацией
      description: Получение статей созданных текущим пользователем с пагинацией
      operationId: getMyArticles
      parameters:
        - in: query
          name: query
          schema:
            type: string
          description: Поисковый запрос по заголовку или содержимому
        - in: query
          name: page
          schema:
            type: integer
            default: 0
            minimum: 0
          description: Номер страницы для пагинации
        - in: query
          name: size
          schema:
            type: integer
            default: 20
            minimum: 1
          description: Размер страницы для пагинации
        - $ref: '#/components/parameters/XUserNameHeader'
      responses:
        '200':
          description: Список статей
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArticlePageResponse'
        '400':
          description: Ошибка параметров запроса
  /api/v1/main_service/articles/{id}:
    get:
      tags:
        - Article
      summary: Получение статьи по ID
      description: Получение статьи по ID
      operationId: getArticle
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Статья
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArticleFullResponse'
        '404':
          description: Статья не найдена

    patch:
      tags:
        - Article
      summary: Обновление статьи по ID
      description: Обновление статьи по ID
      operationId: updateArticle
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        $ref: '#/components/requestBodies/ArticleRequestWithImage'
      responses:
        '200':
          description: Статья успешно обновлена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArticleFullResponse'
        '404':
          description: Статья не найдена
    delete:
      tags:
        - Article
      summary: Удаление статьи по ID
      description: Удаление статьи по ID
      operationId: deleteArticle
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Статья успешно удалена
        '404':
          description: Статья не найдена
...
  /api/v1/main_service/articles/{id}/moderation-status:
    patch:
      tags:
        - Article
      summary: Обновление статуса модерации статьи
      description: Обновление статуса модерации статьи
      operationId: updateArticleModerationPassed
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: boolean
              example: true
      responses:
        '200':
          description: Статус модерации успешно обновлен
        '404':
          description: Статья не найдена
...
  /api/v1/main_service/article-categories:
    get:
      tags:
        - ArticleCategory
      summary: Получение всех категорий статей
      description: Возвращает список всех категорий статей
      operationId: getArticleCategories
      responses:
        '200':
          description: Список категорий статей
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ArticleCategorySummaryResponse'
    post:
      tags:
        - ArticleCategory
      summary: Создание категории статьи
      description: Создание новой категории статьи
      operationId: createArticleCategory
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ArticleCategoryRequest'
      responses:
        '201':
          description: Категория статьи успешно создана
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ArticleCategoryFullResponse'
  /api/v1/main_service/article-categories/{id}:
    get:
      tags:
        - ArticleCategory
      summary: Получение категории статьи по ID
      description: Получение полной информации о категории статьи по ее ID
      operationId: getArticleCategory
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
            minimum: 1
      responses:
        '200':
          description: Категория статьи найдена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArticleCategoryFullResponse'
        '404':
          description: Категория статьи не найдена
          content:
            text/plain:
              schema:
                type: string
    patch:
      tags:
        - ArticleCategory
      summary: Обновление категории статьи
      description: Обновление данных категории статьи по ее ID
      operationId: updateArticleCategory
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
            minimum: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ArticleCategoryRequest'
      responses:
        '200':
          description: Категория статьи успешно обновлена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArticleCategoryFullResponse'
        '404':
          description: Категория статьи не найдена
          content:
            text/plain:
              schema:
                type: string
    delete:
      tags:
        - ArticleCategory
      summary: Удаление категории статьи
      description: Удаление категории статьи по ее ID
      operationId: deleteArticleCategory
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
            minimum: 1
      responses:
        '204':
          description: Категория статьи успешно удалена
        '404':
          description: Категория статьи не найдена
          content:
            text/plain:
              schema:
                type: string
...
components:
  schemas:
...
    ArticlePageResponse:
      allOf:
        - $ref: '#/components/schemas/Page'
        - type: object
          properties:
            content:
              type: array
              items:
                $ref: '#/components/schemas/ArticleSummaryResponse'
...
    ArticleRequest:
      type: object
      description: Запрос на создание или обновление статьи
      properties:
        title:
          type: string
          description: Заголовок статьи
        logo:
          $ref: '#/components/schemas/ImageRequest'
        categoryId:
          type: integer
          format: int64
          description: Категория статьи
        content:
          type: string
          description: Полное содержимое статьи
      required:
        - title
        - content
    ArticleFullResponse:
      type: object
      description: Полный ответ по статье
      properties:
        id:
          type: integer
          format: int64
          description: Уникальный идентификатор статьи
        title:
          type: string
          description: Заголовок статьи
        owner:
          type: string
          description: Владелец/автор статьи
        logoUrl:
          type: string
          format: uri
          description: URL превью статьи
        category:
          type: string
          description: Категория статьи
        content:
          type: string
          description: Полное содержимое статьи
        createdAt:
          type: string
          format: date
          description: Дата публикации статьи
        moderationPassed:
          type: boolean
          description: Флаг, прошла ли статья модерацию
      required:
        - id
        - title
        - owner
        - logoUrl
        - category
        - content
        - createdAt
        - moderationPassed
    ArticleSummaryResponse:
      type: object
      description: Краткая информация о статье
      properties:
        id:
          type: integer
          format: int64
          description: Уникальный идентификатор статьи
        title:
          type: string
          description: Заголовок статьи
        owner:
          type: string
          description: Владелец/автор статьи
        logoUrl:
          type: string
          format: uri
          description: URL превью статьи
        category:
          type: string
          description: Категория статьи
      required:
        - id
        - title
        - owner
        - logoUrl
        - category
    ArticleCategoryRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          example: Technology
    ArticleCategorySummaryResponse:
      type: object
      required:
        - id
        - name
      properties:
        id:
          type: integer
          format: int64
          example: 1
        name:
          type: string
          example: Technology
    ArticleCategoryFullResponse:
      type: object
      required:
        - id
        - name
      properties:
        id:
          type: integer
          format: int64
          example: 1
        name:
          type: string
          example: Technology
...
    Page:
      type: object
      required:
        - content
        - number
        - size
        - totalElements
        - totalPages
      properties:
        content:
          type: array
          description: Массив элементов текущей страницы
          items: { }
        number:
          type: integer
          example: 1
        size:
          type: integer
          example: 1
        totalElements:
          type: integer
          format: int64
          example: 1
        totalPages:
          type: integer
          example: 1
...
    ImageRequest:
      type: string
      format: binary
...
  requestBodies:
    ArticleRequestWithImage:
      required: true
      content:
        multipart/form-data:
          schema:
            type: object
            properties:
              article:
                $ref: '#/components/schemas/ArticleRequest'
              logo:
                $ref: '#/components/schemas/ImageRequest'
...
  parameters:
    XUserNameHeader:
      name: X-User-Name
      in: header
      description: "Имя пользователя, проставляется Gateway"
      required: true
      schema:
        type: string